Working on ccam_folds and what spectra to remove

+++ ccam_normalize.py
@@ -13,7 +13,7 @@ or by the sum for each spectrometer (normtype = 3)
 """
 import numpy
 def ccam_normalize(spectra,wvl,normtype=3):
-    spectra_norm=numpy.zeros(spectra.shape) #create an empty array the same size as spectra
+    spectra_norm=numpy.zeros(spectra.shape,dtype='float32') #create an empty array the same size as spectra
     if normtype==3:
         uv_index=(wvl<=340.797)  #create an index for the UV range
         vis_index=vis_index=(wvl>=382.138)&(wvl<=469.090) #create an index for the VIS range
